+++ CHANGELOG.md
@@ -11,8 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Deprecated
 
 ### Removed
+- 不要なテストコードを削除。
 
 ### Fixed
+- COAの認証エラーが頻出するバグ対応として[sskts-domain](https://www.npmjs.com/package/@motionpicture/sskts-domain)をアップデート。
 
 ### Security